requirements.txt
hydra-core
matplotlib
tqdm
pandas
pre-commit
black
flake8
wandb